Moved NumericalSettings classes to their own module
